feat: Enhance InviteReward dashboard, improve EarnReward list layout, add auto-lock to protection settings, update shortcuts, and adjust WalletListView padding(OK-39501 OK-39529 OK-39514) (#7660)
* Fix hardware sales pending value display(OK-39529)

Replaces the display of 'fiatValue' with 'amount' for pending hardware sales in the InviteReward dashboard. Ensures the correct value is shown to users.

* Add borderRadius to InviteReward dashboard

Applied borderRadius to the YStack and LinearGradient components in the InviteReward dashboard for improved UI consistency and rounded corners.

* Add icons to reward and sales sections in InviteReward

Enhanced the InviteReward dashboard by adding relevant icons next to section titles for cumulative rewards, hardware sales, and on-chain rewards. Also updated the next distribution info box with background and border radius for improved UI consistency.

* Improve layout and responsiveness in EarnReward list

Refactored the EarnReward list item layout to use Stack and flex properties for better responsiveness and alignment, especially on larger screens. Added flexShrink and adjusted flexDirection to enhance text and token display.

* Add auto-lock and unlock for protection settings page(OK-39514)

Introduces an auto-lock mechanism to the protection settings page, locking the UI after 1 minute of inactivity or when the route is unfocused. Adds an unlock button and refactors the PassCodeProtectionSwitch to support onChange callbacks for timer updates.

* Add bottom padding to WalletListView(OK-39501)

Added a contentContainerStyle with paddingBottom to the ListView in WalletListView to improve spacing at the bottom of the list.

* Update shortcut keys and remove copy URL shortcut

Changed the shortcut key for 'explore new tab' to use EShortcutEvents.NewTab2 and removed the 'global_copy_url' shortcut from the sections array in ShortcutsPreview.tsx.