refactor: Refactor settings tab icon order and onPress handlers, add renderTitle prop to Alert component, handle alertItem.alert as string or IEarnText, fix ActionList trigger rendering for async items (OK-39599) (#7663)
* Refactor settings tab icon order and onPress handlers(OK-39599)

Swapped icons for Help Center and Contact Us menu items to better match their functions. Updated all onPress handlers to remove the unused navigation parameter for consistency.

* Add renderTitle prop to Alert component

Introduces a renderTitle prop to the Alert component, allowing custom rendering of the alert title. Updates ProtocolDetailsV2 to use renderTitle for displaying alert titles with EarnText.

* Handle alertItem.alert as string or IEarnText

Updated the AlertSection to support alertItem.alert being either a string or an IEarnText object when rendering EarnText. This improves type safety and flexibility for alert rendering.

* Fix ActionList trigger rendering for async items

Refactors trigger rendering in BasicActionList to ensure the trigger is displayed even when async items are not yet loaded. Uses useMemo for trigger and returns it when renderItemsAsync is true and asyncItems are not available.